requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic==2.6.4
alembic==1.13.1
jinja2==3.1.4
python-multipart==0.0.9
python-jose[cryptography]
google-api-python-client==2.136.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
pwdlib[argon2,bcrypt]>=0.2.1
packaging>=23.2
redis